+++ docs/modules/ROOT/pages/includes/quarkus-temporal.adoc
@@ -8,6 +8,23 @@ h|[.header-title]##Configuration property##
 h|Type
 h|Default
 
+a|icon:lock[title=Fixed at build time] [[quarkus-temporal_quarkus-temporal-netty-allocator-max-order]] [.property-path]##link:#quarkus-temporal_quarkus-temporal-netty-allocator-max-order[`quarkus.temporal.netty.allocator-max-order`]##
+
+[.description]
+--
+The value configuring the `io.grpc.netty.shaded.io.netty.allocator.maxOrder` system property of Netty. The default value is `3`. Configuring this property overrides the minimum `maxOrder` requested by the extensions. This property affects the memory consumption of the application. It must be used carefully. More details on https://programmer.group/pool-area-of-netty-memory-pool.html.
+
+
+ifdef::add-copy-button-to-env-var[]
+Environment variable: env_var_with_copy_button:+++QUARKUS_TEMPORAL_NETTY_ALLOCATOR_MAX_ORDER+++[]
+endif::add-copy-button-to-env-var[]
+ifndef::add-copy-button-to-env-var[]
+Environment variable: `+++QUARKUS_TEMPORAL_NETTY_ALLOCATOR_MAX_ORDER+++`
+endif::add-copy-button-to-env-var[]
+--
+|int
+|
+
 a|icon:lock[title=Fixed at build time] [[quarkus-temporal_quarkus-temporal-enable-mock]] [.property-path]##link:#quarkus-temporal_quarkus-temporal-enable-mock[`quarkus.temporal.enable-mock`]##
 
 [.description]
