+++ extension/runtime/src/main/java/io/quarkiverse/temporal/graal/nettyhandling/BossEventLoopGroup.java
@@ -1,0 +1,23 @@
+package io.quarkiverse.temporal.graal.nettyhandling;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+import jakarta.enterprise.util.AnnotationLiteral;
+import jakarta.inject.Qualifier;
+
+@Qualifier
+@Retention(RetentionPolicy.RUNTIME)
+public @interface BossEventLoopGroup {
+
+    /**
+     * Supports inline instantiation of this qualifier.
+     */
+    public static final class Literal extends AnnotationLiteral<BossEventLoopGroup> implements BossEventLoopGroup {
+
+        public static final Literal INSTANCE = new Literal();
+
+        private static final long serialVersionUID = 1L;
+
+    }
+}
