+++ extension/runtime/src/main/java/io/quarkiverse/temporal/graal/nettyhandling/runtime/virtual/VirtualResponseHandler.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.temporal.graal.nettyhandling.runtime.virtual;
+
+public interface VirtualResponseHandler {
+    void handleMessage(Object msg);
+
+    void close();
+}
