--- extension/runtime/src/main/java/io/quarkiverse/temporal/nettyhandling/runtime/virtual/VirtualResponseHandler.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.temporal.nettyhandling.runtime.virtual;
-
-public interface VirtualResponseHandler {
-    void handleMessage(Object msg);
-
-    void close();
-}
